typing fixes

+++ packages/app/src/app/overmind/namespaces/editor/actions.ts
@@ -63,6 +63,10 @@ export const npmDependencyRemoved: AsyncAction<{
 export const sandboxChanged: AsyncAction<{ id: string }> = withLoadApp<{
   id: string;
 }>(async ({ state, actions, effects }, { id }) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   // This happens when we fork. This can be avoided with state first routing
   if (state.editor.isForkingSandbox) {
     effects.vscode.openModule(state.editor.currentModule);
@@ -171,6 +175,10 @@ export const onOperationApplied: Action<{
   moduleShortid: string;
   code: string;
 }> = ({ state, effects, actions }, { code, moduleShortid }) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   const module = state.editor.currentSandbox.modules.find(
     m => m.shortid === moduleShortid
   );
@@ -197,6 +205,10 @@ export const codeChanged: Action<{
   event?: any;
 }> = ({ effects, state, actions }, { code, event, moduleShortid }) => {
   effects.analytics.trackOnce('Change Code');
+
+  if (!state.editor.currentSandbox) {
+    return;
+  }
 
   const module = state.editor.currentSandbox.modules.find(
     m => m.shortid === moduleShortid
@@ -233,6 +245,10 @@ export const codeChanged: Action<{
 export const saveClicked: AsyncAction = withOwnedSandbox(
   async ({ state, effects, actions }) => {
     const sandbox = state.editor.currentSandbox;
+
+    if (!sandbox) {
+      return;
+    }
 
     try {
       const changedModules = sandbox.modules.filter(module =>
@@ -271,7 +287,7 @@ export const saveClicked: AsyncAction = withOwnedSandbox(
       });
 
       if (
-        state.editor.currentSandbox.originalGit &&
+        sandbox.originalGit &&
         state.workspace.openedWorkspaceItem === 'github'
       ) {
         actions.git.internal.fetchGitChanges();
@@ -288,6 +304,9 @@ export const saveClicked: AsyncAction = withOwnedSandbox(
 );
 
 export const createZipClicked: Action = ({ state, effects }) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
   effects.zip.download(state.editor.currentSandbox);
 };
 
@@ -309,6 +328,10 @@ export const forkSandboxClicked: AsyncAction = async ({
   effects,
   actions,
 }) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   if (
     state.editor.currentSandbox.owned &&
     !state.editor.currentSandbox.customTemplate &&
@@ -347,6 +370,10 @@ export const moduleSelected: Action<{
 
   const sandbox = state.editor.currentSandbox;
 
+  if (!sandbox) {
+    return;
+  }
+
   try {
     let module;
 
@@ -357,9 +384,7 @@ export const moduleSelected: Action<{
         sandbox.directories
       );
     } else {
-      module = state.editor.currentSandbox.modules.find(
-        moduleItem => moduleItem.id === id
-      );
+      module = sandbox.modules.find(moduleItem => moduleItem.id === id);
     }
 
     if (module.shortid === state.editor.currentModuleShortid) {
@@ -368,7 +393,7 @@ export const moduleSelected: Action<{
 
     actions.editor.internal.setCurrentModule(module);
 
-    if (state.live.isLive) {
+    if (state.live.isLive && state.live.liveUser && state.live.roomInfo) {
       effects.vscode.updateUserSelections(
         module,
         actions.live.internal.getSelectionsForModule(module)
@@ -445,6 +470,9 @@ export const prettifyClicked: AsyncAction = async ({
 }) => {
   effects.analytics.track('Prettify Code');
   const module = state.editor.currentModule;
+  if (!module.id) {
+    return;
+  }
   const newCode = await effects.prettyfier.prettify(
     module.id,
     module.title,
@@ -458,13 +486,18 @@ export const prettifyClicked: AsyncAction = async ({
 };
 
 export const errorsCleared: Action = ({ state, effects }) => {
+  const sandbox = state.editor.currentSandbox;
+  if (!sandbox) {
+    return;
+  }
+
   if (state.editor.errors.length) {
     state.editor.errors.forEach(error => {
       try {
         const module = resolveModule(
           error.path,
-          state.editor.currentSandbox.modules,
-          state.editor.currentSandbox.directories
+          sandbox.modules,
+          sandbox.directories
         );
         module.errors = [];
       } catch (e) {
@@ -489,6 +522,10 @@ export const frozenUpdated: AsyncAction<{ frozen: boolean }> = async (
   { state, effects },
   { frozen }
 ) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   state.editor.currentSandbox.isFrozen = frozen;
 
   await effects.api.saveFrozen(state.editor.currentSandbox.id, frozen);
@@ -521,6 +558,10 @@ export const discardModuleChanges: Action<{
   effects.analytics.track('Code Discarded');
 
   const sandbox = state.editor.currentSandbox;
+  if (!sandbox) {
+    return;
+  }
+
   const module = sandbox.modules.find(
     moduleItem => moduleItem.shortid === moduleShortid
   );
@@ -537,6 +578,10 @@ export const fetchEnvironmentVariables: AsyncAction = async ({
   state,
   effects,
 }) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   state.editor.currentSandbox.environmentVariables = await effects.api.getEnvironmentVariables(
     state.editor.currentSandbox.id
   );
@@ -546,6 +591,10 @@ export const updateEnvironmentVariables: AsyncAction<EnvironmentVariable> = asyn
   { state, effects },
   environmentVariable
 ) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   state.editor.currentSandbox.environmentVariables = await effects.api.saveEnvironmentVariable(
     state.editor.currentSandbox.id,
     environmentVariable
@@ -557,6 +606,10 @@ export const updateEnvironmentVariables: AsyncAction<EnvironmentVariable> = asyn
 export const deleteEnvironmentVariable: AsyncAction<{
   name: string;
 }> = async ({ state, effects }, { name }) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   const { id } = state.editor.currentSandbox;
 
   state.editor.currentSandbox.environmentVariables = await effects.api.deleteEnvironmentVariable(
@@ -590,6 +643,9 @@ export const previewActionReceived: Action<any> = (
       });
       break;
     case 'show-error': {
+      if (!state.editor.currentSandbox) {
+        return;
+      }
       const error: ModuleError = {
         column: action.column,
         line: action.line,
@@ -618,6 +674,9 @@ export const previewActionReceived: Action<any> = (
       break;
     }
     case 'show-correction': {
+      if (!state.editor.currentSandbox) {
+        return;
+      }
       const correction: ModuleCorrection = {
         path: action.path,
         column: action.column,
@@ -630,7 +689,7 @@ export const previewActionReceived: Action<any> = (
       };
       try {
         const module = resolveModule(
-          correction.path,
+          correction.path as string,
           state.editor.currentSandbox.modules,
           state.editor.currentSandbox.directories
         );
@@ -644,6 +703,10 @@ export const previewActionReceived: Action<any> = (
       break;
     }
     case 'clear-errors': {
+      const sandbox = state.editor.currentSandbox;
+      if (!sandbox) {
+        return;
+      }
       const currentErrors = state.editor.errors;
 
       const newErrors = clearCorrectionsFromAction(currentErrors, action);
@@ -653,8 +716,8 @@ export const previewActionReceived: Action<any> = (
           try {
             const module = resolveModule(
               error.path,
-              state.editor.currentSandbox.modules,
-              state.editor.currentSandbox.directories
+              sandbox.modules,
+              sandbox.directories
             );
 
             module.errors = [];
@@ -665,8 +728,8 @@ export const previewActionReceived: Action<any> = (
         newErrors.forEach(error => {
           const module = resolveModule(
             error.path,
-            state.editor.currentSandbox.modules,
-            state.editor.currentSandbox.directories
+            sandbox.modules,
+            sandbox.directories
           );
 
           module.errors.push(error);
@@ -677,6 +740,12 @@ export const previewActionReceived: Action<any> = (
       break;
     }
     case 'clear-corrections': {
+      const sandbox = state.editor.currentSandbox;
+
+      if (!sandbox) {
+        return;
+      }
+
       const currentCorrections = state.editor.corrections;
 
       const newCorrections = clearCorrectionsFromAction(
@@ -688,9 +757,9 @@ export const previewActionReceived: Action<any> = (
         state.editor.corrections.forEach(correction => {
           try {
             const module = resolveModule(
-              correction.path,
-              state.editor.currentSandbox.modules,
-              state.editor.currentSandbox.directories
+              correction.path!,
+              sandbox.modules,
+              sandbox.directories
             );
 
             module.corrections = [];
@@ -701,9 +770,9 @@ export const previewActionReceived: Action<any> = (
         });
         newCorrections.forEach(correction => {
           const module = resolveModule(
-            correction.path,
-            state.editor.currentSandbox.modules,
-            state.editor.currentSandbox.directories
+            correction.path!,
+            sandbox.modules,
+            sandbox.directories
           );
 
           module.corrections.push(correction);
@@ -715,6 +784,9 @@ export const previewActionReceived: Action<any> = (
     }
     case 'source.module.rename': {
       const sandbox = state.editor.currentSandbox;
+      if (!sandbox) {
+        return;
+      }
       const module = effects.utils.resolveModule(
         action.path.replace(/^\//, ''),
         sandbox.modules,
@@ -748,6 +820,9 @@ export const renameModule: AsyncAction<{
 }> = withOwnedSandbox(
   async ({ state, actions, effects }, { title, moduleShortid }) => {
     const sandbox = state.editor.currentSandbox;
+    if (!sandbox) {
+      return;
+    }
     const module = sandbox.modules.find(
       moduleItem => moduleItem.shortid === moduleShortid
     );
